feat: 메뉴 삭제

+++ src/main/java/org/example/oristationbackend/service/RestaurantMenuService.java
@@ -7,14 +7,17 @@ import org.example.oristationbackend.dto.restaurant.MenuListResDto;
 import org.example.oristationbackend.dto.restaurant.MenuModReqDto;
 import org.example.oristationbackend.entity.Restaurant;
 import org.example.oristationbackend.entity.RestaurantMenu;
+import org.example.oristationbackend.repository.ReservedMenuRepository;
 import org.example.oristationbackend.repository.RestaurantMenuRepository;
 import org.example.oristationbackend.repository.RestaurantRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.sql.SQLIntegrityConstraintViolationException;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -26,12 +29,14 @@ public class RestaurantMenuService {
   private final RestaurantRepository restaurantRepository;
   private final RestaurantMenuRepository restaurantMenuRepository;
   private final S3Service s3Service;
+  private final ReservedMenuRepository reservedMenuRepository;
 
   @Autowired
-  public RestaurantMenuService(S3Service s3Service, RestaurantRepository restaurantRepository, RestaurantMenuRepository restaurantMenuRepository) {
+  public RestaurantMenuService(S3Service s3Service, RestaurantRepository restaurantRepository, RestaurantMenuRepository restaurantMenuRepository, ReservedMenuRepository reservedMenuRepository) {
     this.s3Service = s3Service;
     this.restaurantRepository = restaurantRepository;
     this.restaurantMenuRepository = restaurantMenuRepository;
+      this.reservedMenuRepository = reservedMenuRepository;
   }
 
   // 식당 id로 식당 메뉴 전체 조회
@@ -79,10 +84,16 @@ public class RestaurantMenuService {
   // 메뉴 id로 메뉴 삭제
   @Transactional(readOnly = false)
   public int deleteRestaurantMenu(int menuId) {
+    if(reservedMenuRepository.existsByRestaurantMenu_MenuId(menuId)){
+      return -1;
+    };
     try {
       restaurantMenuRepository.deleteById(menuId);
-      return 1; // 성공적으로 삭제된 경우 1을 반환
-    } catch (Exception e) {
+      return 1; // 성공적으로 삭제된 경우 1을 반환'
+    }
+    catch(DataIntegrityViolationException e){
+      return -1;
+    }catch (Exception e) {
       return -1; // 에러가 발생한 경우 -1을 반환
     }
   }
